refactor: SolapiUtil의 설정 분리를 위한 SolapiConfig 추가
